refactor :: @Transactional을 detector 레이어에서 service 레이어로 이동
